+++ packages/dso-toolkit/src/components/document-header/document-header.models.ts
@@ -10,4 +10,6 @@ export interface DocumentHeader<TemplateFnReturnType> {
   featureAction?: (e: MouseEvent) => void;
   advancedSelect: AdvancedSelect<unknown>;
   sticky?: boolean;
+  statusMessage?: string;
+  variant?: "vastgesteld" | "ontwerp" | "besluitversie";
 }